pyproject.toml
# Histórico de Alterações:
#   Data       | Desenvolvedor  | Ticket | Observação
#   -----------|----------------|--------|-------------------
#   2026-05-13 | AFSIS Dev      | #001   | Implementação v1.1.0


[project]
name = "fileconverter"
version = "1.0.0"
description = "Conversor universal de arquivos com interface Flet"
requires-python = ">=3.11"
dependencies = [
    "flet>=0.21.0",
    "Pillow>=10.0.0",
    "cairosvg>=2.7.0",
    "PyMuPDF>=1.23.0",
    "pdf2docx>=0.5.8",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
]

[project.scripts]
fileconverter = "src.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100
lint.select = ["E", "F", "I", "N", "W"]
lint.ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
omit = [
    "src/ui/*",
    "src/main.py",
    "src/app.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
